Amazon_cloud_manager
The Platform will provides an Object Oriented language for the
description of Cloud Applications, Services and Resources and it will
perform automatic deployment and interconnection of heterogeneous
resources across heterogeneous cloud systems. Platform allows definition
and automatic connection of monitoring channels for the surveillance of
deployed application components and their operating conditions.